AJout du WS remove dans le code du controller Delete_Container.php

+++ controller/pages/Delete_Container.php
@@ -33,37 +33,64 @@ if($validPage == "DeleteContainer")
         //Récupérer le nom de la machine
         $nomMachine=$machine->getNom();
         //echo "Nom de la machine :".$NomMachine;
-
-        //Récupérer l'id de connexion
-        $connection=Guacamole_ConnectionDAL::findByNom($nomMachine);
-        $connectionId=$connection->getConnectionId();
-        //echo "Id de connexion de la machine :".$connectionId;
-
-        //Supprimer les élements connection_parameter
-        $validDeletePermission=Guacamole_Connection_ParameterDAL::deleteConnection($connectionId);
-        //echo "Suppr Parameter"; 
-
-        //Supprimer les élements connection_permission
-        $validDeleteParameter=Guacamole_Connection_PermissionDAL::deleteConnection($connectionId);
-        //echo "Suppr Permission";      
-
-        //Supprimer le container dans la base guacamole
-        $validDeleteConnection=Guacamole_ConnectionDAL::delete($connectionId);
-        //echo "Suppr Connection";  
-
-        //Suppprimer les partages de cette machine
-        $validDeletePartage=Groupe_has_MachineDAL::deleteMachine($validIdMachine);
-        //echo "Suppr Partage"; 
         
-        //Trouve l'user de la machine et décrémente de 1 son nombre de Container
-        $owner = $machine->getUtilisateur();
-        $owner->setNbVm($owner->getNbVm() - 1); 
+        //=====Appel Web Service pour remove le container sur le serveur de virt=====/
+        try
+        {
+            $client = new SoapCLient(null,
+                array (
+                        'uri' => 'http://virt-server/RemoveContainerRequest',
+                        'location' => 'http://virt-server/remove_container_ws.php',
+                        'trace' => 1,
+                        'exceptions' => 0
+                ));
+            $result = $client->__call('removeContainer',
+                array(
+                        'nameContainer' => $nomMachine
+                ));
+        }
+        catch(SoapFault $f)
+        {
+            echo $f;
+        }
 
-        //Supprimer le container dans la base DBVirtDemand
-        $validDeleteMachine=MachineDAL::delete($validIdMachine);
-        //echo "final";
+        if ($result == "0")
+        {
+            //Récupérer l'id de connexion
+            $connection=Guacamole_ConnectionDAL::findByNom($nomMachine);
+            $connectionId=$connection->getConnectionId();
+            //echo "Id de connexion de la machine :".$connectionId;
+
+            //Supprimer les élements connection_parameter
+            $validDeletePermission=Guacamole_Connection_ParameterDAL::deleteConnection($connectionId);
+            //echo "Suppr Parameter"; 
+
+            //Supprimer les élements connection_permission
+            $validDeleteParameter=Guacamole_Connection_PermissionDAL::deleteConnection($connectionId);
+            //echo "Suppr Permission";      
+
+            //Supprimer le container dans la base guacamole
+            $validDeleteConnection=Guacamole_ConnectionDAL::delete($connectionId);
+            //echo "Suppr Connection";  
+
+            //Suppprimer les partages de cette machine
+            $validDeletePartage=Groupe_has_MachineDAL::deleteMachine($validIdMachine);
+            //echo "Suppr Partage"; 
         
-        $message="ok";
+            //Trouve l'user de la machine et décrémente de 1 son nombre de Container
+            $owner = $machine->getUtilisateur();
+            $owner->setNbVm($owner->getNbVm() - 1); 
+
+            //Supprimer le container dans la base DBVirtDemand
+            $validDeleteMachine=MachineDAL::delete($validIdMachine);
+            //echo "final";
+        
+            $message="ok";
+        }
+        else
+        {
+            echo "Erreur lors de la suppression du container nommé ". $nomMachine; //TODO log
+        }
     }
 }
